+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 Compatibility:
 
 * Implemented `String#start_with?(Regexp)` (#1771).
+Bug fixes:
+
+* Dump and load instance variables in subclasses of `Exception` (#1766).
 
 # 19.3.0